requirements.txt
uvicorn>=0.34.2
pydantic>=2.11.4
starlette>=0.46.2
fastapi>=0.115.12
joblib>=1.4.2
retry_requests>=2.0.0
requests_cache>= 1.2.1
openmeteo_requests>=1.4.0
tensorflow>=2.19.0
pandas>=1.5.0
xgboost>=1.6.0
matplotlib>=3.5.0
seaborn>=0.11.0
cdsapi>=0.7.5
xarray>=2022.0.0
dotenv>=0.9.9
netcdf4>=1.6.0
numpy>=1.20.0
scikit-learn>=1.1.0
scipy>=1.9.0
dask>=2022.0.0
psutil>=5.9.0
pyarrow>=10.0.0
pathlib>=1.0.1 
seaborn>=0.13.2
openpyxl>=3.1.5
geopy>=2.4.1